+++ src/talk2powersystemllm/app/server/healthchecks/__init__.py
@@ -1,7 +1,11 @@
 from .graphdb_healthcheck import GraphDBHealthchecker
 from .healthchecks import HealthChecks
+from .llm_healthcheck import LLMHealthchecker
+from .redis_healthcheck import RedisHealthchecker
 
 __all__ = [
     "HealthChecks",
     "GraphDBHealthchecker",
+    "RedisHealthchecker",
+    "LLMHealthchecker",
 ]