+++ tests/acceptance_tests/test_app.py
@@ -1,12 +1,13 @@
 import random
 import re
+import time
 from datetime import datetime
 from unittest import TestCase
 
 import requests
 from requests import Response
 
-from .openai_mock import mock_openai, mock_openai_verify, mock_openai_reset
+from .openai_mock import mock_openai, mock_openai_verify, mock_openai_reset, mock_openai_error
 
 
 def is_valid_uuid_v4(s: str) -> bool:
@@ -30,9 +31,11 @@ class AcceptanceTestsApp(TestCase):
     def setUp(self):
         super(AcceptanceTestsApp, self).setUp()
         mock_openai_reset()
+        self.llm_healthcheck_is_ok()
 
     def tearDown(self):
         super(AcceptanceTestsApp, self).tearDown()
+        self.llm_healthcheck_is_ok()
         mock_openai_verify()
 
     def validate_response_and_return_json_body(
@@ -65,7 +68,7 @@ class AcceptanceTestsApp(TestCase):
         self.assertTrue("status" in actual_response_json)
         self.assertEqual("OK", actual_response_json["status"])
         self.assertTrue("healthChecks" in actual_response_json)
-        self.assertEqual(2, len(actual_response_json["healthChecks"]))
+        self.assertEqual(3, len(actual_response_json["healthChecks"]))
         self.assertTrue(
             {
                 "status": "OK",
@@ -95,6 +98,19 @@ class AcceptanceTestsApp(TestCase):
                 "troubleshooting": "http://talk2powersystem:8000/__trouble#redis-health-check-status-is-not-ok",
                 "description": "Checks if Redis can be queried.",
                 "message": "Redis can be queried."
+            } in actual_response_json["healthChecks"]
+        )
+        self.assertTrue(
+            {
+                "status": "OK",
+                "severity": "HIGH",
+                "id": "http://talk2powersystem.no/talk2powersystem-api/llm-healthcheck",
+                "name": "LLM Health Check",
+                "type": "llm",
+                "impact": "Some requests to the chat bot failed during the last 60 seconds due to LLM errors!",
+                "troubleshooting": "http://talk2powersystem:8000/__trouble#llm-health-check-status-is-not-ok",
+                "description": "Checks if any LLM calls resulted in errors during the last 60 seconds!",
+                "message": "No LLM errors were hit in the last 60 seconds!"
             } in actual_response_json["healthChecks"]
         )
 
@@ -104,7 +120,7 @@ class AcceptanceTestsApp(TestCase):
         self.assertTrue("status" in actual_response_json)
         self.assertEqual("OK", actual_response_json["status"])
         self.assertTrue("healthChecks" in actual_response_json)
-        self.assertEqual(2, len(actual_response_json["healthChecks"]))
+        self.assertEqual(3, len(actual_response_json["healthChecks"]))
         self.assertTrue(
             {
                 "status": "OK",
@@ -134,6 +150,19 @@ class AcceptanceTestsApp(TestCase):
                 "troubleshooting": "http://talk2powersystem:8000/__trouble#redis-health-check-status-is-not-ok",
                 "description": "Checks if Redis can be queried.",
                 "message": "Redis can be queried."
+            } in actual_response_json["healthChecks"]
+        )
+        self.assertTrue(
+            {
+                "status": "OK",
+                "severity": "HIGH",
+                "id": "http://talk2powersystem.no/talk2powersystem-api/llm-healthcheck",
+                "name": "LLM Health Check",
+                "type": "llm",
+                "impact": "Some requests to the chat bot failed during the last 60 seconds due to LLM errors!",
+                "troubleshooting": "http://talk2powersystem:8000/__trouble#llm-health-check-status-is-not-ok",
+                "description": "Checks if any LLM calls resulted in errors during the last 60 seconds!",
+                "message": "No LLM errors were hit in the last 60 seconds!"
             } in actual_response_json["healthChecks"]
         )
 
@@ -624,7 +653,8 @@ SELECT ?link ?name ?format ?description ?kind {{
                     "type": "function",
                     "function": {
                         "name": "sparql_query",
-                        "arguments": "{ \"query\": \"SELECT * { ?substation a <https://cim.ucaiug.io/ns#Substation> }\" }"
+                        "arguments": "{ \"query\": \"SELECT * { ?substation a <https://cim.ucaiug.io/ns#Substation> "
+                                     "}\" }"
                     }
                 }
             ],
@@ -649,7 +679,8 @@ SELECT ?link ?name ?format ?description ?kind {{
                             "id": "sparql_query_call_1",
                             "function": {
                                 "name": "sparql_query",
-                                "arguments": "{\"query\": \"SELECT * { ?substation a <https://cim.ucaiug.io/ns#Substation> }\"}"
+                                "arguments": "{\"query\": \"SELECT * { ?substation a "
+                                             "<https://cim.ucaiug.io/ns#Substation> }\"}"
                             }
                         }
                     ]
@@ -688,7 +719,8 @@ SELECT ?link ?name ?format ?description ?kind {{
                             "id": "sparql_query_call_1",
                             "function": {
                                 "name": "sparql_query",
-                                "arguments": "{\"query\": \"SELECT * { ?substation a <https://cim.ucaiug.io/ns#Substation> }\"}"
+                                "arguments": "{\"query\": \"SELECT * { ?substation a "
+                                             "<https://cim.ucaiug.io/ns#Substation> }\"}"
                             }
                         }
                     ]
@@ -739,7 +771,8 @@ SELECT ?link ?name ?format ?description ?kind {{
                             "id": "sparql_query_call_1",
                             "function": {
                                 "name": "sparql_query",
-                                "arguments": "{\"query\": \"SELECT * { ?substation a <https://cim.ucaiug.io/ns#Substation> }\"}"
+                                "arguments": "{\"query\": \"SELECT * { ?substation a "
+                                             "<https://cim.ucaiug.io/ns#Substation> }\"}"
                             }
                         }
                     ]
@@ -832,4 +865,92 @@ SELECT ?link ?name ?format ?description ?kind {{
                 ]
             },
             explain_response_body2
+        )
+
+    def test_llm_errors_openai_401_error(self) -> None:
+        question = "list substations"
+        mock_openai_error(
+            request_messages=[
+                {
+                    "content": question,
+                    "role": "user"
+                }
+            ],
+            status_code=401,
+            error_message="Access denied due to invalid subscription key or wrong API endpoint. Make sure "
+                          "to provide a valid key for an active subscription and use a correct regional "
+                          "API endpoint for your resource.",
+            n_times=1
+        )
+
+        conversation_request_body = {"question": question}
+        conversation_response = requests.post(CONVERSATION_ENDPOINT, json=conversation_request_body, timeout=(2, 10))
+        self.assertEqual(500, conversation_response.status_code)
+
+        self.llm_healthcheck_warning()
+        time.sleep(61)
+
+    def test_llm_errors_openai_429_error(self) -> None:
+        question = "list substations"
+        mock_openai_error(
+            request_messages=[
+                {
+                    "content": question,
+                    "role": "user"
+                }
+            ],
+            status_code=429,
+            error_message="You exceeded your current quota, please check your plan and billing details. "
+                          "For more information on this error, read the docs: "
+                          "https://platform.openai.com/docs/guides/error-codes/api-errors.",
+            n_times=3
+        )
+
+        conversation_request_body = {"question": question}
+        conversation_response = requests.post(CONVERSATION_ENDPOINT, json=conversation_request_body, timeout=(2, 10))
+        self.assertEqual(500, conversation_response.status_code)
+
+        self.llm_healthcheck_warning()
+        time.sleep(61)
+
+    def llm_healthcheck_is_ok(self):
+        response = requests.get(HEALTH_ENDPOINT, timeout=(2, 10))
+        actual_response_json = self.validate_response_and_return_json_body(response)
+        self.assertEqual(2, len(actual_response_json))
+        self.assertTrue("status" in actual_response_json)
+        self.assertEqual("OK", actual_response_json["status"])
+        self.assertTrue("healthChecks" in actual_response_json)
+        self.assertTrue(
+            {
+                "status": "OK",
+                "severity": "HIGH",
+                "id": "http://talk2powersystem.no/talk2powersystem-api/llm-healthcheck",
+                "name": "LLM Health Check",
+                "type": "llm",
+                "impact": "Some requests to the chat bot failed during the last 60 seconds due to LLM errors!",
+                "troubleshooting": "http://talk2powersystem:8000/__trouble#llm-health-check-status-is-not-ok",
+                "description": "Checks if any LLM calls resulted in errors during the last 60 seconds!",
+                "message": "No LLM errors were hit in the last 60 seconds!"
+            } in actual_response_json["healthChecks"]
+        )
+
+    def llm_healthcheck_warning(self):
+        response = requests.get(HEALTH_ENDPOINT, timeout=(2, 10))
+        actual_response_json = self.validate_response_and_return_json_body(response)
+        self.assertEqual(2, len(actual_response_json))
+        self.assertTrue("status" in actual_response_json)
+        self.assertEqual("WARNING", actual_response_json["status"])
+        self.assertTrue("healthChecks" in actual_response_json)
+        self.assertTrue(
+            {
+                "status": "WARNING",
+                "severity": "HIGH",
+                "id": "http://talk2powersystem.no/talk2powersystem-api/llm-healthcheck",
+                "name": "LLM Health Check",
+                "type": "llm",
+                "impact": "Some requests to the chat bot failed during the last 60 seconds due to LLM errors!",
+                "troubleshooting": "http://talk2powersystem:8000/__trouble#llm-health-check-status-is-not-ok",
+                "description": "Checks if any LLM calls resulted in errors during the last 60 seconds!",
+                "message": "LLM errors were hit in the last 60 seconds!"
+            } in actual_response_json["healthChecks"]
         )